+++ src/Symfony/Component/Debug/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 2.6.0
 -----
 
+* generalized ErrorHandler and ExceptionHandler,
+  with some new methods and others deprecated
 * enhanced error messages for uncaught exceptions
 
 2.5.0